fix(ci): move brand test helper into src
