+++ apps/medusa-be/tests/unit/src/workflows/brand.unit.spec.ts
@@ -1,11 +1,11 @@
-import { describe, expect, it } from "vitest"
 import {
   createScope,
   ensureProductsAssignableToBrand,
   getBrandProductsLockKeys,
   getProductBrandIdsToReplace,
   getProductBrandLockKeys,
-} from "../helpers/mocks"
+} from "src/test-helpers/mocks"
+import { describe, expect, it } from "vitest"
 
 describe("brand workflow helpers", () => {
   describe("product brand lock keys", () => {
